+++ packages/cypher-builder/src/Pattern.ts
@@ -19,7 +19,7 @@
 
 import { stringifyObject } from "./utils/stringify-object";
 import { escapeLabel } from "./utils/escape-label";
-import { padLeft } from "./utils/utils";
+import { padLeft } from "./utils/pad-left";
 import type { NodeRef } from "./references/NodeRef";
 import type { RelationshipRef } from "./references/RelationshipRef";
 import type { CypherEnvironment } from "./Environment";
